streamlitpages/smartdecision/requirements.txt
streamlit==1.50.0
numpy>=1.26.0
scipy>=1.11.0
python-dotenv>=1.0.1
google-genai==1.42.0
